Check ensemble permissions

+++ src/handlers/ensemble.ts
@@ -3,6 +3,8 @@ import { MyContext } from "../context";
 import { Ensemble } from "@prisma/client";
 import { ensembleMenu } from "../menus/ensembleMenu";
 import { deleteEnsemble } from "../utils/models/ensemble";
+import { telegramUserIsMember } from "../utils/models/membership";
+import { telegramUserIsAdmin } from "../utils/models/admin";
 
 export const createEnsembleHandler: MiddlewareFn<MyContext> = async (ctx) => {
   await ctx.reply(ctx.t("create_command_answer"), {
@@ -14,6 +16,15 @@ export const createEnsembleHandler: MiddlewareFn<MyContext> = async (ctx) => {
 export const printEnsembleHandler: (
   ensemble: Ensemble
 ) => MiddlewareFn<MyContext> = (ensemble) => async (ctx) => {
+  if (
+    !(await telegramUserIsMember({
+      telegramUserId: ctx.from?.id,
+      ensembleId: ensemble.id,
+    }))
+  ) {
+    ctx.reply("No eres miembro de esta agrupación.");
+    return;
+  }
   ctx.session.ensembleId = ensemble.id;
   await ctx.reply(ctx.templates.ensembleDetailTemplate({ ensemble }), {
     parse_mode: "HTML",
@@ -24,6 +35,15 @@ export const printEnsembleHandler: (
 export const deleteEnsembleHandler: (
   ensemble: Ensemble
 ) => MiddlewareFn<MyContext> = (ensemble) => async (ctx) => {
+  if (
+    !(await telegramUserIsAdmin({
+      telegramUserId: ctx.from?.id,
+      ensembleId: ensemble.id,
+    }))
+  ) {
+    ctx.reply("No eres administrador de esta agrupación.");
+    return;
+  }
   await deleteEnsemble({ ensembleId: ensemble.id });
   await ctx.reply(`La agrupación "${ensemble.name}" ha sido eliminada.`);
 };
